fix(db): SQL-aware migration splitter + optimistic locking on tasks
Closes Gemini's two outstanding correctness concerns from #845:

1. PgMigrator::apply split each migration's SQL on a literal `;`. Any
   future migration that embeds a semicolon in a string literal,
   PL/pgSQL function body, or dollar-quoted block would crash
   mid-migration. Replaces the naive split with `pg_split_statements`,
   which walks the SQL with a small state machine that recognizes:
   single-quoted strings (with `''` escape), double-quoted identifiers,
   `--` line comments, `/* */` block comments, and `$tag$ ... $tag$`
   dollar-quoted blocks. Backed by 13 unit tests.

2. The `tasks.version` column existed (migration v18) but no UPDATE used
   it as a precondition. The primary write path (`TaskDb::update`) now
   reads the expected version from the loaded `TaskState`, sends
   `WHERE id = $N AND version = $M`, increments atomically inside the
   same UPDATE, and surfaces a typed conflict error when
   rows_affected == 0. The in-memory `TaskStore` cache mirrors the
   increment after every successful persist / claim / release so the
   next write sends a matching expected version.

Out of scope per the issue's narrowing comment: JSONB column
migration. `exec_plans.data` is already JSONB (#861); the remaining
TEXT-encoded JSON columns on `tasks` are intentionally left as TEXT
in this PR.

Refactors driven by the 800-line per-file ceiling:
- `db_pg.rs`: tests and the new splitter live in sibling files
  (`db_pg_tests.rs`, `db_pg_split.rs`).
- `task_db/queries_tasks.rs`: startup-recovery write paths
  (`apply_replayed_state`, `recover_in_progress`) move to
  `queries_recovery.rs`.

Closes #845
