+++ crates/harness-core/src/lib.rs
@@ -4,6 +4,7 @@ pub mod capability;
 pub mod config;
 pub mod db;
 pub mod db_pg;
+mod db_pg_split;
 pub mod error;
 pub mod interceptor;
 pub mod lang_detect;
