+++ crates/harness-server/src/task_db/queries_aux.rs
@@ -142,7 +142,7 @@ impl TaskDb {
             "SELECT t.id, t.task_kind, t.status, t.turn, t.pr_url, t.rounds, t.error, t.source, \
                     t.external_id, t.parent_id, t.created_at, t.updated_at, t.repo, t.depends_on, \
                     t.project, t.priority, t.phase, t.description, t.request_settings, \
-                    t.scheduler_state, \
+                    t.scheduler_state, t.version, \
                     c.triage_output, c.plan_output, c.pr_url AS ck_pr_url, \
                     c.last_phase, \
                     TO_CHAR(c.updated_at AT TIME ZONE 'UTC', 'YYYY-MM-DD\"T\"HH24:MI:SS\"Z\"') AS ck_updated_at \
@@ -183,6 +183,7 @@ impl TaskDb {
                 description: row.description,
                 request_settings: row.request_settings,
                 scheduler_state: row.scheduler_state,
+                version: row.version,
             };
             let task_state = match task_row.try_into_task_state() {
                 Ok(s) => s,
